[dbt] br_me_comex_stat

+++ models/br_me_comex_stat/municipio_exportacao.sql
@@ -20,6 +20,6 @@ SAFE_CAST(id_sh4 AS STRING) id_sh4,
 SAFE_CAST(id_pais AS STRING) id_pais,
 SAFE_CAST(sigla_uf AS STRING) sigla_uf,
 SAFE_CAST(id_municipio AS STRING) id_municipio,
-SAFE_CAST(peso_liquido_kg AS INT64) peso_liquido_kg,
-SAFE_CAST(valor_fob_dolar AS INT64) valor_fob_dolar
+SAFE_CAST(peso_liquido_kg AS FLOAT64) peso_liquido_kg,
+SAFE_CAST(valor_fob_dolar AS FLOAT64) valor_fob_dolar
 FROM basedosdados-dev.br_me_comex_stat_staging.municipio_exportacao AS t
